requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
PyYAML==6.0.2
paramiko==3.5.1
